generic-git-operator/requirements.txt
kopf==1.34.0
pyyaml==6.0
GitPython==3.1.24
requests==2.26.0
kubernetes
paramiko